+++ packages/react-wallet/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @lightsparkdev/react-wallet
 
+## 0.5.0
+
+### Minor Changes
+
+- Require signing on the initialize wallet operation to allow channels to be created.
+
+### Patch Changes
+
+- Updated dependencies
+  - @lightsparkdev/wallet-sdk@0.5.0
+
+## 0.4.1
+
+### Patch Changes
+
+- Fix a bug in the request_withrawal mutation.
+- Updated dependencies
+  - @lightsparkdev/wallet-sdk@0.4.1
+
 ## 0.4.0
 
 ### Minor Changes